refactor: extract constants and file icon helpers (#22)
- Create Constants.swift for Docker, Keychain, GitHub, UI values
- Create FileIconHelper.swift for file icon/color/language mapping
- Remove duplicate code from EditorView and FileExplorerView
- All 22 tests passing
