+++ Sources/Zero/Constants.swift
@@ -1,0 +1,24 @@
+import Foundation
+
+enum Constants {
+    enum Docker {
+        static let path = "/usr/local/bin/docker"
+        static let baseImage = "alpine:latest"
+        static let workspacePath = "/workspace"
+    }
+    
+    enum Keychain {
+        static let service = "com.zero.ide"
+        static let account = "github_token"
+    }
+    
+    enum GitHub {
+        static let pageSize = 30
+    }
+    
+    enum UI {
+        static let sidebarMinWidth: CGFloat = 220
+        static let sidebarIdealWidth: CGFloat = 260
+        static let sidebarMaxWidth: CGFloat = 400
+    }
+}
